+++ src/core/primitives/index.ts
@@ -1,1 +1,4 @@
+export * from './JCanvas';
+export * from './JInput';
 export * from './JLabel';
+export * from './JParagraph';
