+++ src/pages/solidText/page/style.scss
